requirements.txt
beautifulsoup4==4.10.0
Flask==2.0.1
psycopg2==2.9.3
python-dotenv==0.20.0
requests==2.25.1
apscheduler==3.9.1
